refact : VoteResultRepository 반환 자료형 충돌문제 해결
totalVoteCount -> Long 형으로
CZ=365

+++ src/main/java/com/example/manymanyUsers/statistics/controller/StatisticsController.java
@@ -26,7 +26,7 @@ public class StatisticsController {
     @GetMapping("/vote/{voteId}/total-statistics")
     public ResponseEntity<TotalStatisticsResponse> getTotalStatistics(@PathVariable("voteId") Long voteId) {
 
-        int totalVoteNumber = statisticsService.getTotalStatistics(voteId);
+        Long totalVoteNumber = statisticsService.getTotalStatistics(voteId);
 
         TotalStatisticsResponse wholeStatisticsResponse = TotalStatisticsResponse.builder()
                 .voteId(voteId)
