+++ src/main/java/com/example/manymanyUsers/statistics/dto/TotalStatisticsResponse.java
@@ -10,11 +10,11 @@ import lombok.NoArgsConstructor;
 public class TotalStatisticsResponse {
 
     private Long voteId;
-    private int totalVote;
+    private Long totalVote;
     private String message;
 
     @Builder
-    public TotalStatisticsResponse(Long voteId, int totalVote, String message) {
+    public TotalStatisticsResponse(Long voteId, Long totalVote, String message) {
         this.voteId = voteId;
         this.totalVote = totalVote;
         this.message = message;
